@@ -35,8 +35,8 @@ class Post extends Model
      *
      * @var array
      */
-    protected $dates = [
-        'posted_at'
+    protected $casts = [
+        'posted_at' => 'datetime'
     ];
 
     /**
