@@ -9,10 +9,8 @@ class DevDatabaseSeeder extends Seeder
 {
     /**
      * Run the database seeds.
-     *
-     * @return void
      */
-    public function run()
+    public function run(): void
     {
         Post::factory()
             ->count(20)
