@@ -95,11 +95,8 @@ class CommentTest extends TestCase
 
     /**
      * Valid params for updating or creating a resource
-     *
-     * @param  array  $overrides new params
-     * @return array Valid params for updating or creating a resource
      */
-    private function validParams($overrides = [])
+    private function validParams(array $overrides = []): array
     {
         $post = Post::factory()->create();
 
