@@ -75,11 +75,8 @@ class UserTest extends TestCase
 
     /**
      * Valid params for updating or creating a resource
-     *
-     * @param  array  $overrides new params
-     * @return array  Valid params for updating or creating a resource
      */
-    private function validParams($overrides = [])
+    private function validParams(array $overrides = []): array
     {
         return array_merge([
             'name' => 'Anakin',
